fix: address QA round 3 — discount math, invalid codes, subscribe crash, opacity
- C1: Add CSS override to ensure page content is always visible even when
  framer-motion animations don't complete (tunnel/throttled environments)
- C5: Fix discount calculation to use dollars (not cents), fix SAVE10 seed
  value from 1000 to 10
- C6: Check isValid flag from backend response, show error for invalid codes
- S1: Include nested Plan object in SubscriptionDto response so frontend
  can access plan limits; add optional chaining as safety net
- Fix ValidateDiscountResponse field names to match backend (isValid, not valid)
- Fix ValidateDiscountResponse price types from long to decimal

+++ backend/AcademixLMS.Application/DTOs/Discount/DiscountDtos.cs
@@ -65,6 +65,6 @@ public class ValidateDiscountResponse
     public string? Message { get; set; }
     public string? DiscountType { get; set; }
     public decimal? DiscountValue { get; set; }
-    public long? OriginalPrice { get; set; }
-    public long? DiscountedPrice { get; set; }
+    public decimal? OriginalPrice { get; set; }
+    public decimal? DiscountedPrice { get; set; }
 }
